task delete fix
fix + error handling

+++ WorkBoard/Controllers/TaskController.cs
@@ -61,10 +61,20 @@ namespace WorkBoard.Controllers
 
         public IActionResult Delete(int id)
         {
-            Task task = db.Tasks.Find(id);
-            db.Tasks.Remove(task);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            
+            if(ModelState.IsValid)
+            {
+                Task task = db.Tasks.Find(id);
+                if (task == null)
+                {
+                    return NotFound();
+                }
+                db.Tasks.Remove(task);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            return View();
+            
         }
 
         public IActionResult Edit(int id)
